[US24] Implement account deletion on the server side

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/repository/UserRepository.java
@@ -16,5 +16,5 @@ public interface UserRepository extends JpaRepository<ApplicationUser, Long> {
     Page<ApplicationUser> findAll(Pageable pageable);
 
 
-    List<ApplicationUser> findAllByPointsGreaterThan(int number);
+    List<ApplicationUser> findAllByPointsGreaterThan(Long number);
 }
